flask_backend/requirements_core.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
Werkzeug==2.3.7

# Core scraping dependencies (Python 3.13 compatible)
telethon==1.29.3
cryptg==0.4.0
instagrapi==2.0.0
requests-oauthlib==1.3.1
instaloader>=4.10

# Web scraping utilities
beautifulsoup4==4.12.2
lxml==4.9.3
fake-useragent==1.4.0

# Async support
aiohttp==3.8.6

# Job scheduling
APScheduler==3.10.4

# Image processing (core only)
Pillow==10.0.1

# Logging
structlog==23.1.0

# Database (optional - can use SQLite instead)
# psycopg2-binary==2.9.7  # Uncomment for PostgreSQL support
